refactor(wallet): extract BalanceCard, StickyBalancePill, TokensList, ActivityFeed components
